rename class viewController
