+++ hnh1/AuthentificationViewController.swift
@@ -1,42 +1,24 @@
 //
-//  AuthentificationViewController.swift
+//  ViewController.swift
 //  hnh1
 //
-//  Created by Юлия Руднева on 14.04.2022.
-//  CustomButton1.swift
+//  Created by Юлия Руднева on 05.04.2022.
+//
 
 import UIKit
 
-class AuthentificationViewController: UIViewController, UITextFieldDelegate {
-
-    override func viewDidLoad() {
-        super.viewDidLoad()
-        let tap = UITapGestureRecognizer (target: self, action: #selector (endEditing))
-        view.addGestureRecognizer(tap)
-        loginTextField.delegate = self
-        passwordTextField.delegate = self
-        passwordConfirmTextField.delegate = self
-    }
+class AuthentificationViewController: UIViewController {
     
     @IBOutlet weak var loginTextField: UITextField!
     @IBOutlet weak var passwordTextField: UITextField!
-    @IBOutlet weak var passwordConfirmTextField: UITextField!
-    @IBOutlet weak var doneButton: flickeringButton!
+    @IBOutlet weak var loginButton: UIButton!
+    @IBOutlet weak var registerButton: UIButton!
     
-    func textFieldShouldReturn(_ textField: UITextField) -> Bool {
-        if textField == loginTextField {
-            passwordTextField.becomeFirstResponder()
-        }
-        else if textField == passwordTextField {
-            passwordConfirmTextField.becomeFirstResponder()
-        }
-        else {
-            textField.resignFirstResponder()
-        }
-        return true
+    override func viewDidLoad() {
+        super.viewDidLoad()
+        // Do any additional setup after loading the view.
+        
     }
 
-    @objc func endEditing() {
-        view.endEditing(true)
-    }
 }
+
