+++ hnh1/Base.lproj/Main.storyboard
@@ -9,10 +9,10 @@
         <capability name="documents saved in the Xcode 8 format" minToolsVersion="8.0"/>
     </dependencies>
     <scenes>
-        <!--Authentification View Controller-->
+        <!--Registration View Controller-->
         <scene sceneID="qq7-GB-Bk0">
             <objects>
-                <viewController storyboardIdentifier="AuthentificationViewController" id="dZC-gO-MJN" customClass="AuthentificationViewController" customModule="hnh1" customModuleProvider="target" sceneMemberID="viewController">
+                <viewController storyboardIdentifier="AuthentificationViewController" id="dZC-gO-MJN" customClass="RegistrationViewController" customModule="hnh1" customModuleProvider="target" sceneMemberID="viewController">
                     <view key="view" contentMode="scaleToFill" id="LiX-K7-uCY">
                         <rect key="frame" x="0.0" y="0.0" width="414" height="896"/>
                         <autoresizingMask key="autoresizingMask" widthSizable="YES" heightSizable="YES"/>
@@ -109,10 +109,10 @@
             </objects>
             <point key="canvasLocation" x="19" y="-8"/>
         </scene>
-        <!--View Controller-->
+        <!--Authentification View Controller-->
         <scene sceneID="tne-QT-ifu">
             <objects>
-                <viewController id="BYZ-38-t0r" customClass="ViewController" customModule="hnh1" customModuleProvider="target" sceneMemberID="viewController">
+                <viewController id="BYZ-38-t0r" customClass="AuthentificationViewController" customModule="hnh1" customModuleProvider="target" sceneMemberID="viewController">
                     <view key="view" contentMode="scaleToFill" id="8bC-Xf-vdC">
                         <rect key="frame" x="0.0" y="0.0" width="414" height="896"/>
                         <autoresizingMask key="autoresizingMask" widthSizable="YES" heightSizable="YES"/>
