--- hnh1/ViewController.swift
@@ -1,24 +1,0 @@
-//
-//  ViewController.swift
-//  hnh1
-//
-//  Created by Юлия Руднева on 05.04.2022.
-//
-
-import UIKit
-
-class ViewController: UIViewController {
-    
-    @IBOutlet weak var loginTextField: UITextField!
-    @IBOutlet weak var passwordTextField: UITextField!
-    @IBOutlet weak var loginButton: UIButton!
-    @IBOutlet weak var registerButton: UIButton!
-    
-    override func viewDidLoad() {
-        super.viewDidLoad()
-        // Do any additional setup after loading the view.
-        
-    }
-
-}
-
